demo2/requirements.txt
# 环境变量管理
python-dotenv==1.0.0

# 终端UI
rich>=13.0.0

# 其他依赖
pydantic>=2.0.0

# 向量数据库
chromadb>=0.4.0

# HTTP 请求
requests>=2.31.0

# 智谱AI SDK
zai-sdk>=0.2.0
